LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/month-expected.txt

Inputs Month


FAIL month type support on input element assert_equals: expected "month" but got "text"
FAIL The value attribute, if specified and not empty, must have a value that is a valid month string assert_equals: expected "2011-01" but got "1999-01"
FAIL The min attribute, if specified, must have a value that is a valid month string. assert_equals: expected "" but got "invalid_min"
FAIL The max attribute, if specified, must have a value that is a valid month string assert_equals: expected "2099-01" but got "2011-12"
FAIL User agents must not allow the user to set the value to a non-empty string that is not a valid month string. assert_equals: expected "" but got "invalid-month"
PASS Month value can be empty string. 
FAIL When value attribute has two digits year value, the value,which is invalid, must return empty string. assert_equals: expected "" but got "13-06"
FAIL When value is set with invalid value, the value must return empty string. assert_equals: expected "" but got "invalid value"
FAIL When value is given invalid value to non-empty valid string, the value must be same as before. assert_equals: expected "2013-05" but got "invalid value"
FAIL When step attribute is given invalid value, it must ignore the invalid value and use defaul value instead. The object is in an invalid state.
FAIL Month should be <= 13: If the value of the element is not a valid month string, then set it to the empty string instead. assert_equals: expected "" but got "2013-13"
FAIL Month should be > 0: If the value of the element is not a valid month string, then set it to the empty string instead.> assert_equals: expected "" but got "2013-00"
